backend/requirements.txt
fastapi
uvicorn
boto3
opensearch-py
strands-agents
pydantic
python-multipart
twelvelabs
requests
